add assets now allows clicking on the whole row
